+++ .gitignore
@@ -15,4 +15,4 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
+.idea/*
