+++ src/css/App.css
@@ -3,12 +3,13 @@
 }
 
 .App-logo {
+
   animation: App-logo-spin infinite 20s linear;
   height: 80px;
 }
 
 .App-header {
-  background-color: #222;
+  background-color: #FFFFFF;
   height: 150px;
   padding: 20px;
   color: white;
